replace context-provide-plugin to separate dir. separate context-provide-plugin-component
