+++ src/vue-book/book-main.js
@@ -4,7 +4,7 @@ import VueClipboard from 'vue-clipboard2'
 import Router from 'vue-router'
 import { VueBookComponents, createRoute } from 'vue-book'
 import { ColorThemePlugin } from '../services/ColorThemePlugin'
-import { ContextProvidePlugin } from '../services/ContextProvidePlugin'
+import { ContextProvidePlugin } from '../services/context-provide/ContextProvidePlugin'
 import { BusPlugin } from 'vue-epic-bus'
 import { registerVuesticObject } from '../components/resize-events'
 import { DropdownPopperPlugin } from '../components/vuestic-components/va-dropdown/dropdown-popover-subplugin'
